[ANDROSDK-526] Rewrite SectionChildrenAppender using SingleParentChildStore
