+++ core/src/main/java/org/hisp/dhis/android/core/dataset/SectionHandler.java
@@ -30,6 +30,7 @@ package org.hisp.dhis.android.core.dataset;
 import org.hisp.dhis.android.core.arch.handlers.IdentifiableSyncHandlerImpl;
 import org.hisp.dhis.android.core.arch.handlers.SyncHandler;
 import org.hisp.dhis.android.core.common.HandleAction;
+import org.hisp.dhis.android.core.common.IdentifiableObjectStore;
 import org.hisp.dhis.android.core.common.LinkModelHandler;
 import org.hisp.dhis.android.core.common.LinkModelHandlerImpl;
 import org.hisp.dhis.android.core.common.ObjectWithUid;
@@ -46,7 +47,7 @@ public class SectionHandler extends IdentifiableSyncHandlerImpl<Section> {
     private final SyncHandler<DataElementOperand> greyedFieldsHandler;
     private final LinkModelHandler<DataElementOperand, SectionGreyedFieldsLinkModel> sectionGreyedFieldsLinkHandler;
 
-    SectionHandler(SectionStore sectionStore,
+    SectionHandler(IdentifiableObjectStore<Section> sectionStore,
                    OrderedLinkModelHandler<ObjectWithUid, SectionDataElementLinkModel> sectionDataElementLinkHandler,
                    SyncHandler<DataElementOperand> greyedFieldsHandler,
                    LinkModelHandler<DataElementOperand, SectionGreyedFieldsLinkModel> sectionGreyedFieldsLinkHandler) {
@@ -74,7 +75,7 @@ public class SectionHandler extends IdentifiableSyncHandlerImpl<Section> {
 
     public static SectionHandler create(DatabaseAdapter databaseAdapter) {
         return new SectionHandler(
-                SectionStoreImpl.create(databaseAdapter),
+                SectionStore.create(databaseAdapter),
                 new OrderedLinkModelHandlerImpl<ObjectWithUid,
                         SectionDataElementLinkModel>(SectionDataElementLinkStore.create(databaseAdapter)),
                 DataElementOperandHandler.create(databaseAdapter),
