+++ core/src/main/java/org/hisp/dhis/android/core/dataset/SectionStore.java
@@ -28,11 +28,48 @@
 
 package org.hisp.dhis.android.core.dataset;
 
+import android.database.Cursor;
+import android.database.sqlite.SQLiteStatement;
+import android.support.annotation.NonNull;
 
+import org.hisp.dhis.android.core.arch.db.binders.IdentifiableStatementBinder;
+import org.hisp.dhis.android.core.arch.db.binders.StatementBinder;
+import org.hisp.dhis.android.core.arch.db.tableinfos.SingleParentChildProjection;
+import org.hisp.dhis.android.core.common.CursorModelFactory;
 import org.hisp.dhis.android.core.common.IdentifiableObjectStore;
+import org.hisp.dhis.android.core.common.StoreFactory;
+import org.hisp.dhis.android.core.common.UidsHelper;
+import org.hisp.dhis.android.core.data.database.DatabaseAdapter;
 
-import java.util.List;
+import static org.hisp.dhis.android.core.utils.StoreUtils.sqLiteBind;
 
-interface SectionStore extends IdentifiableObjectStore<Section> {
-    List<Section> getDataSet(String dataSetUid);
+public final class SectionStore {
+
+    private SectionStore() {}
+
+    private static StatementBinder<Section> BINDER = new IdentifiableStatementBinder<Section>() {
+        @Override
+        public void bindToStatement(@NonNull Section o, @NonNull SQLiteStatement sqLiteStatement) {
+            super.bindToStatement(o, sqLiteStatement);
+            sqLiteBind(sqLiteStatement, 7, o.description());
+            sqLiteBind(sqLiteStatement, 8, o.sortOrder());
+            sqLiteBind(sqLiteStatement, 9, UidsHelper.getUidOrNull(o.dataSet()));
+            sqLiteBind(sqLiteStatement, 10, o.showRowTotals());
+            sqLiteBind(sqLiteStatement, 11, o.showColumnTotals());
+        }
+    };
+
+    static final CursorModelFactory<Section> FACTORY = new CursorModelFactory<Section>() {
+        @Override
+        public Section fromCursor(Cursor cursor) {
+            return Section.create(cursor);
+        }
+    };
+
+    static final SingleParentChildProjection CHILD_PROJECTION = new SingleParentChildProjection(
+            SectionTableInfo.TABLE_INFO, SectionFields.DATA_SET);
+
+    public static IdentifiableObjectStore<Section> create(DatabaseAdapter databaseAdapter) {
+        return StoreFactory.objectWithUidStore(databaseAdapter, SectionTableInfo.TABLE_INFO, BINDER, FACTORY);
+    }
 }
