--- core/src/main/java/org/hisp/dhis/android/core/dataset/SectionStoreImpl.java
@@ -1,90 +1,0 @@
-/*
- * Copyright (c) 2017, University of Oslo
- *
- * All rights reserved.
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- * Redistributions of source code must retain the above copyright notice, this
- * list of conditions and the following disclaimer.
- *
- * Redistributions in binary form must reproduce the above copyright notice,
- * this list of conditions and the following disclaimer in the documentation
- * and/or other materials provided with the distribution.
- * Neither the name of the HISP project nor the names of its contributors may
- * be used to endorse or promote products derived from this software without
- * specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
- * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-
-package org.hisp.dhis.android.core.dataset;
-
-import android.database.Cursor;
-import android.database.sqlite.SQLiteStatement;
-import android.support.annotation.NonNull;
-
-import org.hisp.dhis.android.core.arch.db.WhereClauseBuilder;
-import org.hisp.dhis.android.core.arch.db.binders.IdentifiableStatementBinder;
-import org.hisp.dhis.android.core.arch.db.binders.StatementBinder;
-import org.hisp.dhis.android.core.common.CursorModelFactory;
-import org.hisp.dhis.android.core.common.IdentifiableObjectStoreImpl;
-import org.hisp.dhis.android.core.common.SQLStatementBuilder;
-import org.hisp.dhis.android.core.common.SQLStatementWrapper;
-import org.hisp.dhis.android.core.common.UidsHelper;
-import org.hisp.dhis.android.core.data.database.DatabaseAdapter;
-
-import java.util.List;
-
-import static org.hisp.dhis.android.core.utils.StoreUtils.sqLiteBind;
-
-public final class SectionStoreImpl extends IdentifiableObjectStoreImpl<Section>
-        implements SectionStore {
-
-    private SectionStoreImpl(DatabaseAdapter databaseAdapter,
-                                         SQLStatementWrapper statementWrapper,
-                                         SQLStatementBuilder statementBuilder) {
-        super(databaseAdapter, statementWrapper, statementBuilder, BINDER, FACTORY);
-    }
-
-    private static StatementBinder<Section> BINDER = new IdentifiableStatementBinder<Section>() {
-        @Override
-        public void bindToStatement(@NonNull Section o, @NonNull SQLiteStatement sqLiteStatement) {
-            super.bindToStatement(o, sqLiteStatement);
-            sqLiteBind(sqLiteStatement, 7, o.description());
-            sqLiteBind(sqLiteStatement, 8, o.sortOrder());
-            sqLiteBind(sqLiteStatement, 9, UidsHelper.getUidOrNull(o.dataSet()));
-            sqLiteBind(sqLiteStatement, 10, o.showRowTotals());
-            sqLiteBind(sqLiteStatement, 11, o.showColumnTotals());
-        }
-    };
-
-    private static final CursorModelFactory<Section> FACTORY = new CursorModelFactory<Section>() {
-        @Override
-        public Section fromCursor(Cursor cursor) {
-            return Section.create(cursor);
-        }
-    };
-
-    @Override
-    public List<Section> getDataSet(String dataSetUid) {
-        String whereClause = new WhereClauseBuilder()
-                .appendKeyStringValue(SectionFields.DATA_SET, dataSetUid)
-                .build();
-        return selectWhereClauseAsList(whereClause);
-    }
-
-    public static SectionStore create(DatabaseAdapter databaseAdapter) {
-        SQLStatementBuilder statementBuilder = new SQLStatementBuilder(SectionTableInfo.TABLE_INFO);
-        SQLStatementWrapper statementWrapper = new SQLStatementWrapper(statementBuilder, databaseAdapter);
-        return new SectionStoreImpl(databaseAdapter, statementWrapper, statementBuilder);
-    }
-}
